+++ js/app.config.js
@@ -1,20 +1,29 @@
 app.config = {
 	env					: 'prod',
-	version				: '1.8.59', // Error handeling for username, adjusted stroke_width on track creation, added time for track creation
+	version				: '1.8.62', // multiple download added to communtiy, added community model.
 	disconnect_timeout_to_stop_polling: 45000, // stop trying to find tables after 45 seconds
 	envs	: {
 		alpha: {	// loads local data only **5 tap on No Table Found Screen to launch shell app
 			base_url	: 'http://app.dev.withease.io:3001/', // local
-			api_url		: 'https://api.sisyphus.withease.io/',
-			web_url		: 'http://dev.webcenter.sisyphus-industries.com/',
+			api_url		: 'https://webcenter.sisyphus-industries.com/',
+			web_url		: 'https://webcenter.sisyphus-industries.com/',
 			sisbot_url  : 'http://api.dev.withease.io:3000/', //talking to sisbot
 			port		: 3001,
 		},
 		beta: {		// tests local network
 			base_url	: 'http://app.dev.withease.io:3001/', //local url
-			api_url		: 'http://localhost:3000/', // add entry in your computers /etc/hosts mapped to your bot's IP address
-			web_url		: 'http://localhost:3000/', //web_center url	***Change to this for Rails web_center= http://localhost:3000/  (aka rails s) //  10.0.0.3	beta_bot.local
-			sisbot_url  : 'http://192.168.86.20:3002', //talking to sisbot    //  ... or just put your URL in here '192.168.XX.XXX:3002' << for local Dev Env --insert your ip address + 3000
+			api_url		: 'http://localhost:3333/', // add entry in your computers /etc/hosts mapped to your bot's IP address
+			web_url		: 'http://localhost:3333/', //web_center url	***Change to this for Rails web_center= http://localhost:3333/  (aka rails s) //  10.0.0.3	beta_bot.local
+			sisbot_url  : 'http://192.168.1.6:3002', //talking to sisbot    //  ... or just put your URL in here '192.168.XX.XXX:3002' << for local Dev Env --insert your ip address + 3000
+			port		: 3001,
+		},
+		matt: {
+			base_url	: 'http://app.dev.withease.io:3001/', //local url
+			api_url		: 'https://webcenter.sisyphus-industries.com/',
+			web_url		: 'https://webcenter.sisyphus-industries.com/',
+			// api_url		: 'http://localhost:3333/', // add entry in your computers /etc/hosts mapped to your bot's IP address
+			// web_url		: 'http://localhost:3333/', //web_center url	***Change to this for Rails web_center= http://localhost:3333/  (aka rails s) //  10.0.0.3	beta_bot.local
+			sisbot_url  : 'http://192.168.86.206:3002', //talking to sisbot    //  ... or just put your URL in here '192.168.XX.XXX:3002' << for local Dev Env --insert your ip address + 3000
 			port		: 3001,
 		},
 		sisbot: {
@@ -49,10 +58,10 @@ app.config = {
 		return this.envs[this.env].sisbot_url;
 	},
 	set_sisbot_url: function (data) {
-		console.log("Sisbot URL: ", data);
+		console.warn("Sisbot URL: ", data);
 		if(!data.match(/^https?:\/\//i)) data = 'http://' + data;
 
-		if(!data.match(/:[0-9]+\/?$/i)) data += ":3002/";
+		if(!data.match(/:[0-9]+\/?$/i)) data += ":3002";
 		this.envs[this.env].sisbot_url = data;
 	},
 	get_webcenter_url: function () {
@@ -77,7 +86,23 @@ app.config = {
 
 // if its an ip address or sisyphus.local, it'll set itself to sisbot
 if (window.location.href.indexOf('withease') < 0)			app.config.env = 'sisbot';
-if (window.location.href.indexOf('localhost') > -1)			app.config.env = 'beta';
+if (window.location.href.indexOf('localhost') > -1) {
+	var fragment = window.location.hash;
+	var env = 'beta';
+
+	if (fragment.charAt(0) === '#') {
+		fragment = fragment.slice(1);
+		fragment = fragment.split('&');
+		var count		= fragment.length;
+		for (var i = 0; i < count; i++) {
+			var obj = fragment[i].split('=');
+			if (obj[0] == 'env') env = obj[1];
+		}
+	}
+
+	console.log("Env:", env);
+	app.config.env = env;
+}
 if (window.location.href.indexOf('.local') > -1)        	app.config.env = 'sisbot';
 if (window.location.href.indexOf('192.168') > -1) 			app.config.env = 'sisbot';